requirements.txt
<<<<<<< HEAD
streamlit
google-generativeai
python-dotenv
=======
streamlit
google-generativeai
python-dotenv
>>>>>>> 24bf564 (⚡ First commit of Flash 2.0 with layout and logo)
